+++ frontends/api/src/hooks/channels/index.ts
@@ -3,7 +3,7 @@ import {
   useMutation,
   useQueryClient,
 } from "@tanstack/react-query"
-import { useQuery } from "../../useQueryCacheWarning"
+import { useQuery } from "../../ssr/useQuery"
 import { channelsApi } from "../../clients"
 import type {
   ChannelsApiChannelsListRequest,
