+++ frontends/main/src/app/departments/page.tsx
@@ -2,23 +2,20 @@ import React from "react"
 import { Metadata } from "next"
 import DepartmentListingPage from "@/app-pages/DepartmentListingPage/DepartmentListingPage"
 import { standardizeMetadata } from "@/common/metadata"
-import { QueryClient, dehydrate, Hydrate } from "@tanstack/react-query"
+import { Hydrate } from "@tanstack/react-query"
 import { learningResourcesKeyFactory } from "api/hooks/learningResources"
 import { channelsKeyFactory } from "api/hooks/channels"
+import { prefetch } from "api/ssr/prefetch"
 
 export const metadata: Metadata = standardizeMetadata({
   title: "Departments",
 })
 
 const Page: React.FC = async () => {
-  const queryClient = new QueryClient()
-
-  await Promise.all([
-    queryClient.prefetchQuery(channelsKeyFactory.countsByType("department")),
-    queryClient.prefetchQuery(learningResourcesKeyFactory.schools()),
+  const dehydratedState = await prefetch([
+    channelsKeyFactory.countsByType("department"),
+    learningResourcesKeyFactory.schools(),
   ])
-
-  const dehydratedState = dehydrate(queryClient)
 
   return (
     <Hydrate state={dehydratedState}>
