+++ frontends/main/src/app/providers.tsx
@@ -1,26 +1,17 @@
 "use client"
 
-import React, { useEffect } from "react"
+import React from "react"
 import { getQueryClient } from "./getQueryClient"
 import { QueryClientProvider } from "@tanstack/react-query"
 import { ThemeProvider, NextJsAppRouterCacheProvider } from "ol-components"
 import { Provider as NiceModalProvider } from "@ebay/nice-modal-react"
 import ConfiguredPostHogProvider from "@/components/ConfiguredPostHogProvider/ConfiguredPostHogProvider"
+import { useQueryCacheWarning } from "api/ssr/useQueryCacheWarning"
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   const queryClient = getQueryClient()
 
-  useEffect(() => {
-    const onLoad = () => {
-      queryClient.setQueryData(["initialRenderComplete"], true)
-      window.removeEventListener("load", onLoad)
-    }
-    if (document.readyState === "complete") {
-      onLoad()
-    } else {
-      window.addEventListener("load", onLoad)
-    }
-  })
+  useQueryCacheWarning(queryClient)
 
   return (
     <ConfiguredPostHogProvider>
